attempt at new algo
this algo tries to find the best multi pair by pushing through a loop, comparing old mutlipliers and the diff of the multi pair. if both progress negatively (increase), it breaks the loop, finding the first most efficient pair. problem: the best pair would be 5 * 5 + 1, this algo couldn't find it.